(PC-31124)[API] feat: add command to reindex all indexed offers
