+++ api/src/pcapi/scripts/install.py
@@ -28,11 +28,11 @@ def install_commands(app: flask.Flask) -> None:
         "pcapi.scripts.booking.commands",
         "pcapi.scripts.check_pre_migrations",
         "pcapi.scripts.external_users.commands",
-        "pcapi.scripts.full_index_offers",
         "pcapi.scripts.install_data",
         "pcapi.scripts.generate_expected_openapi_json",
         "pcapi.scripts.provider.check_provider_api",
         "pcapi.scripts.sandbox",
+        "pcapi.scripts.search.full_index_offers",
         "pcapi.scripts.ubble_archive_past_identifications",
         "pcapi.scripts.offer.fix_offer_data_titelive",
         "pcapi.scripts.offer.fix_product_gtl_id_titelive",
