+++ api/src/pcapi/scripts/search/full_index_collective_offers.py
